+++ Core/src/core/types.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct CoreCapabilities {
     pub schema_version: u32,
@@ -14,26 +14,26 @@ pub struct CoreCapabilities {
     pub localization_tables: Vec<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct ChainSummary {
     pub chain_name: String,
     pub curve: String,
     pub default_network: Option<String>,
     pub default_derivation_path: Option<String>,
-    pub endpoint_count: usize,
-    pub settings_visible_endpoint_count: usize,
-    pub explorer_endpoint_count: usize,
+    pub endpoint_count: u64,
+    pub settings_visible_endpoint_count: u64,
+    pub explorer_endpoint_count: u64,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct LocalizationSummary {
     pub supported_locales: Vec<String>,
     pub tables: Vec<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct CoreBootstrap {
     pub capabilities: CoreCapabilities,
